reorganize config files
