+++ Sources/client-sdk-swift/config/TopicConfiguration.swift
@@ -1,5 +1,13 @@
 import Foundation
 
+protocol TopicClientConfigurationProtocol {
+    var loggerFactory: MomentoLoggerFactoryProtocol { get }
+    var transportStrategy: TransportStrategyProtocol { get }
+    
+    func withTransportStrategy(transportStrategy: TransportStrategyProtocol) -> TopicClientConfigurationProtocol
+    func withClientTimeout(timeout: TimeInterval) -> TopicClientConfigurationProtocol
+}
+
 public class TopicClientConfiguration: TopicClientConfigurationProtocol {
     var loggerFactory: MomentoLoggerFactoryProtocol
     var transportStrategy: TransportStrategyProtocol
