--- Sources/client-sdk-swift/config/TopicConfigurationProtocol.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-protocol TopicClientConfigurationProtocol {
-    var loggerFactory: MomentoLoggerFactoryProtocol { get }
-    var transportStrategy: TransportStrategyProtocol { get }
-    
-    func withTransportStrategy(transportStrategy: TransportStrategyProtocol) -> TopicClientConfigurationProtocol
-    func withClientTimeout(timeout: TimeInterval) -> TopicClientConfigurationProtocol
-}
