+++ Sources/client-sdk-swift/config/TopicConfigurations.swift
@@ -1,7 +1,7 @@
-let defaultLoggerFactory = DefaultMomentoLoggerFactory()
+// Prebuilt configurations for Momento Topics clients
 
 public class Default: TopicClientConfiguration {
     static func latest() -> TopicClientConfiguration {
-        return TopicClientConfiguration(loggerFactory: defaultLoggerFactory, transportStrategy: StaticTransportStrategy(grpcConfig: StaticGrpcConfiguration(deadline: 15.0)))
+        return TopicClientConfiguration(loggerFactory: DefaultMomentoLoggerFactory(), transportStrategy: StaticTransportStrategy(grpcConfig: StaticGrpcConfiguration(deadline: 15.0)))
     }
 }
