+++ Sources/client-sdk-swift/config/transport/GrpcConfiguration.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+protocol GrpcConfigurationProtocol {
+    var deadline: TimeInterval { get }
+    static func withDeadline(deadline: TimeInterval) -> GrpcConfigurationProtocol
+}
+
+public class StaticGrpcConfiguration: GrpcConfigurationProtocol {
+    var deadline: TimeInterval
+    
+    init(deadline: TimeInterval) {
+        self.deadline = deadline
+    }
+    
+    static func withDeadline(deadline: TimeInterval) -> GrpcConfigurationProtocol {
+        return StaticGrpcConfiguration(deadline: deadline)
+    }
+}
