--- Sources/client-sdk-swift/config/transport/GrpcConfigurationProtocol.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol GrpcConfigurationProtocol {
-    var deadline: TimeInterval { get }
-    static func withDeadline(deadline: TimeInterval) -> GrpcConfigurationProtocol
-}
